Remove  bunch of unnecessary casts and change a bunch of calls to static methods to use the appropriate static class

+++ rt/ws/security/src/main/java/org/apache/cxf/ws/security/tokenstore/EHCacheTokenStore.java
@@ -141,7 +141,7 @@ public class EHCacheTokenStore implements TokenStore, Closeable, BusLifeCycleLis
 
     private int getTTL() {
         int parsedTTL = (int)ttl;
-        if (ttl != (long)parsedTTL) {
+        if (ttl != parsedTTL) {
              // Fall back to 60 minutes if the default TTL is set incorrectly
             parsedTTL = 3600;
         }
